requirements.txt
siaskynet==2.3.0
telegram==0.0.1
python-dotenv==0.19.0
